pyproject.toml
[tool.poetry]
name = "tvtime-scrapy-api"
version = "0.1.0"
description = "Unofficial TVTime API via Scrapy"
authors = ["giraycoskun <giraycoskun@sabanciuniv.edu>"]
license = "MIT"
readme = "README.md"
packages = [{include = "src"}]


[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.1"
Scrapy = "^2.8.0"
uvicorn = "^0.21.1"
python-dotenv = "^1.0.0"
loguru = "^0.7.0"
celery = {extras = ["redis"], version = "^5.2.7"}
redis-om = "^0.1.2"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
python-jose = "^3.3.0"
sqlalchemy = "^2.0.12"
psycopg2 = "^2.9.6"


[tool.poetry.group.docs.dependencies]
mkdocs-material = "^9.1.8"
mkdocstrings = "^0.21.2"
mkdocstrings-python = "^0.9.0"
mkdocs-swagger-ui-tag = "^0.6.1"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"
black = "^23.3.0"
pre-commit = "^3.3.1"
pylint = "^2.17.4"


[tool.poetry.group.test.dependencies]
pytest = "^7.3.1"
pytest-asyncio = "^0.21.0"
httpx = "^0.24.0"
coverage = "<7.0"
coveralls = "^3.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
